CLN: Move all rest tests to one location
